Move lock impls out of lib.rs
